+++ tests/integration/bench_tests.rs
@@ -2,6 +2,8 @@
 
 use deno_core::url::Url;
 use test_util as util;
+use test_util::itest;
+use test_util::itest_flaky;
 use util::assert_contains;
 use util::assert_not_contains;
 use util::env_vars_for_npm_tests;
