+++ tests/integration/coverage_tests.rs
@@ -3,6 +3,7 @@
 use deno_core::serde_json;
 use std::fs;
 use test_util as util;
+use test_util::itest;
 use test_util::TempDir;
 use util::assert_starts_with;
 use util::env_vars_for_npm_tests;
