+++ tests/integration/js_unit_tests.rs
@@ -1,4 +1,5 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
 use std::io::BufRead;
 use std::io::BufReader;
 use std::time::Duration;
