+++ tests/integration/jsr_tests.rs
@@ -3,6 +3,7 @@
 use deno_core::serde_json::Value;
 use deno_lockfile::Lockfile;
 use test_util as util;
+use test_util::itest;
 use url::Url;
 use util::env_vars_for_jsr_tests;
 use util::TestContextBuilder;
