+++ tests/integration/jupyter_tests.rs
@@ -1,5 +1,7 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
+use test_util::itest;
+
 itest!(jupyter_install_command_not_exists {
   args: "jupyter --install",
   output: "jupyter/install_command_not_exists.out",
