+++ tests/integration/mod.rs
@@ -1,90 +1,5 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
-#[macro_export]
-macro_rules! itest(
-($name:ident {$( $key:ident: $value:expr,)*})  => {
-  #[test]
-  fn $name() {
-    let test = test_util::CheckOutputIntegrationTest {
-      $(
-        $key: $value,
-       )*
-      .. Default::default()
-    };
-    let output = test.output();
-    output.assert_exit_code(test.exit_code);
-    if !test.output.is_empty() {
-      assert!(test.output_str.is_none());
-      output.assert_matches_file(test.output);
-    } else {
-      output.assert_matches_text(test.output_str.unwrap_or(""));
-    }
-  }
-}
-);
-
-#[macro_export]
-macro_rules! itest_flaky(
-($name:ident {$( $key:ident: $value:expr,)*})  => {
-  #[flaky_test::flaky_test]
-  fn $name() {
-    let test = test_util::CheckOutputIntegrationTest {
-      $(
-        $key: $value,
-       )*
-      .. Default::default()
-    };
-    let output = test.output();
-    output.assert_exit_code(test.exit_code);
-    if !test.output.is_empty() {
-      assert!(test.output_str.is_none());
-      output.assert_matches_file(test.output);
-    } else {
-      output.assert_matches_text(test.output_str.unwrap_or(""));
-    }
-  }
-}
-);
-
-#[macro_export]
-macro_rules! context(
-({$( $key:ident: $value:expr,)*})  => {
-  test_util::TestContext::create(test_util::TestContextOptions {
-    $(
-      $key: $value,
-      )*
-    .. Default::default()
-  })
-}
-);
-
-#[macro_export]
-macro_rules! itest_steps(
-($name:ident {$( $key:ident: $value:expr,)*})  => {
-  #[test]
-  fn $name() {
-    (test_util::CheckOutputIntegrationTestSteps {
-      $(
-        $key: $value,
-       )*
-      .. Default::default()
-    }).run()
-  }
-}
-);
-
-#[macro_export]
-macro_rules! command_step(
-({$( $key:ident: $value:expr,)*})  => {
-  test_util::CheckOutputIntegrationTestCommandStep {
-    $(
-      $key: $value,
-      )*
-    .. Default::default()
-  }
-}
-);
-
 // These files have `_tests.rs` suffix to make it easier to tell which file is
 // the test (ex. `lint_tests.rs`) and which is the implementation (ex. `lint.rs`)
 // when both are open, especially for two tabs in VS Code
