+++ tests/integration/npm_tests.rs
@@ -5,6 +5,7 @@ use deno_core::serde_json::json;
 use deno_core::serde_json::Value;
 use pretty_assertions::assert_eq;
 use test_util as util;
+use test_util::itest;
 use util::assert_contains;
 use util::env_vars_for_npm_tests;
 use util::http_server;
