+++ tests/integration/publish_tests.rs
@@ -5,6 +5,7 @@ use test_util::assert_contains;
 use test_util::assert_not_contains;
 use test_util::env_vars_for_jsr_tests;
 use test_util::env_vars_for_npm_tests;
+use test_util::itest;
 use test_util::TestContextBuilder;
 
 itest!(no_token {
