+++ tests/integration/vendor_tests.rs
@@ -6,6 +6,7 @@ use pretty_assertions::assert_eq;
 use std::fmt::Write as _;
 use std::path::PathBuf;
 use test_util as util;
+use test_util::itest;
 use test_util::TempDir;
 use util::http_server;
 use util::new_deno_dir;
