+++ tests/integration/worker_tests.rs
@@ -1,5 +1,7 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 
+use test_util::itest;
+
 itest!(worker_error {
   args: "run -A workers/worker_error.ts",
   output: "workers/worker_error.ts.out",
