--- tests/integration_tests.rs
@@ -1,7 +1,0 @@
-// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-
-// The tests exist in a sub folder instead of as separate files in
-// this directory so that cargo doesn't compile each file as a new crate.
-
-#[cfg(test)]
-mod integration;
